+++ odbc-api/tests/common.rs
@@ -38,7 +38,7 @@ pub fn setup_empty_table(
 
 pub fn cursor_to_string(cursor: impl Cursor) -> String {
     let batch_size = 20;
-    let mut buffer = buffers::TextRowSet::for_cursor(batch_size, &cursor).unwrap();
+    let mut buffer = buffers::TextRowSet::for_cursor(batch_size, &cursor, None).unwrap();
     let mut row_set_cursor = cursor.bind_buffer(&mut buffer).unwrap();
 
     let mut text = String::new();